typing test

+++ matbert_ner/models/bert_model.py
@@ -91,6 +91,7 @@ class BERTNER(BertPreTrainedModel):
         sequence_output = outputs[0]
         # valid outputs
         sequence_output, label_ids, attention_mask = valid_sequence_output(sequence_output, label_ids, attention_mask, valid_mask, device)
+        # label_ids = label_ids.type(torch.long)
         # dropout on valid hidden layer output
         sequence_output = self.dropout(sequence_output)
         # classification logits
@@ -99,7 +100,7 @@ class BERTNER(BertPreTrainedModel):
         prediction_ids = self.crf.decode(logits, mask=attention_mask)
         # if labels are provided, calculate loss
         if label_ids is not None:
-            loss = -self.crf(logits, label_ids.type(torch.long), mask=attention_mask)
+            loss = -self.crf(logits, label_ids, mask=attention_mask)
         # return statements
         if return_logits and label_ids is not None:
             return loss, logits, prediction_ids
